requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
argostranslate==1.9.6
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
ctranslate2==4.3.1
emoji==2.12.1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
magic-filter==1.0.12
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==2.0.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
packaging==24.1
protobuf==5.27.2
pydantic==2.7.4
pydantic_core==2.18.4
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
sacremoses==0.0.53
sentencepiece==0.2.0
setuptools==70.2.0
six==1.16.0
stanza==1.1.1
sympy==1.12.1
torch==2.3.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
mtranslate~=1.8